+++ cli/bin/form_concierge_cli.dart
@@ -1,0 +1,8 @@
+import 'dart:io';
+
+import 'package:form_concierge_cli/form_concierge_cli.dart';
+
+Future<void> main(List<String> args) async {
+  final code = await runFormConciergeCli(args);
+  exit(code);
+}