+++ examples/flutter_mobile_full/lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:form_concierge_client/form_concierge_client.dart';
-import 'package:form_concierge_survey_widget/form_concierge_survey_widget.dart';
+import 'package:form_concierge/form_concierge.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 const _apiUrl = String.fromEnvironment(